Monorepo refactor
TLDR; Pnpm is in the root of the repo. Which forces
all PNPM projects to use workspaces. This moves 
everything into a subfolder so a separate Pnpm project
outside of workspaces can be used.

---------------------------------------
The mobile apps will use React Native. However this
currently does not support PNPM workspaces. It
does work in a flat PNPM project though.

This refactor moves all the current projects into a
'gg-mono' folder. Now that the pnpm-workspaces
file is not in the root a new react native project
can be added.

PNPM workspace support looks to be coming in
early 2024.