Auto deploy android build to google play store (#100)
What triggers a build and release to the Play store?
A merge to 'main' and updated Android release notes.

Is it fully automatic for a release?
Not right now, but only because alpha and beta tracks
have not been setup on Google Play. These require the
app to be reviewed and right now that app doesn't do 
enough to be submitted.

For now after a merge to main a manual release has to
be completed on the Play Store.

How long does a build take?
A full build without a cache currently takes around 4
minutes. However with cache it is half of that time.