Tweak sonarqube setup (#254)
This removes the sonar scan step from the github-action.
It was taking a lot of time and the only reason for this
local run was to run the code coverage tool. However
this was never setup fully so was just a pointless, slow,
extra step.

This PR also makes the sonar project names simpler.