Publish assetlink with multiple fingerprints (#281)
Apps published from the playstore have the production fingerprint. That cannot be used locally
without taking over the app signing completely which is something to avoid. So to be able to test local
builds the debug.keystore fingerprint needs to be added to the assetlinks.json.

The Google docs are not crystal clear here. So this is the first attempt to add the debug fingerprint.
https://developer.android.com/training/app-links/verify-android-applinks