Ensure all builds work (#340)
iOS, Android and Web all now correctly auth and work.
The Xcode cloud ci script no longer needs a workaround
for a broken version of Cocoapods. 1.15.2 fixed the issue.
The script was also given a tidy to use the CI variables
for the paths to the project instead of hardcoding them.