Optomise being 'authenticated' (#513)
Step one was that after getting a valid access/refresh token from the bungie web code the app did an additional call to get a valid token from the end point that consumes the previous valid token. This additional call is not needed.

The second step is being authenticated only happened when there was a valid token AND a valid bungie account.
Now just having the token is enough. However the base view now needs to show the user states such as trying to
get the account. If this fails an error should be shown and then logout should be called.