feat: The 3 characters inventory should always match up after scrolling
When on an inventory page e.g. armor the expected behaviour when you horizontal scroll to another guardian is that the items align. So if you were on 'helmets' scrolling left or right shows the other characters helmets. As each character has an independent list, this code keeps that illusion by tracking the scrolloffset.y and then moving each guardian list to the correct position.