fix: Some icons were missing season overlay (#847)
Root cause: there was code that returned false e.g. if (index), but the index was 0. A legitamte value for an index. So the code has been updated to instead be `if (index !== undefined). These are for values extracted from JSON so if there is no value at all, it will be undefined.