refactor(transfer): Simplify the system
This adds a 'previousCharacterId' property to allow the UI to update. Without this the moveItem store setter would not know where an item has moved between. which would have meant transferItem() or the acual move function would have to track and manager that.

The zustand states are also slightly simplified by better use of the 'mutative' helper library.