refactor(getProfile): Simplify validation but keep complex type (#1187)
The Valibot validation of getProfile takes 100ms of a high end laptop. It's not strictly necessary as in most cases the profile is either fine or totally broken. So a simple validation for the main sections is enough for validation. Then the full type can be applied to it for autocomplete.