refactor: break up inventory data
Before the different pages were all grouped together. For example weaponsInventory was a single object in the state. This meant if one item changed all 4 pages were updated.

The new system breaks the pages to each character (guardians and vault). It then does an eqaulity check to ensure only characters with changed data are updated.

Previously transfers could cause the UI to update and take up to 1 second to redraw everything. It seems the largest updates are now 250ms.