refactor: Optomise how page data is updated (#1380)
Each inventory page is now a seperate array in the store that can be updated on its own. As opposed to a large ggCharacters object having to be updated regardless of which page changed.

Then instead of always creating a costly mutative 'create' object the checks for new data are done first. If the data changes then then mutative is used.