refactor: Update the bottom sheet (#1601)
This change stops the entire view being a full 'modal'. A bottom sheet should never have been this type of modal and it caused touch to be unresponsive on iOS for a second after the sheet slide off.

Touch events would also fire on button in a scrollview when they were touched, but the view was scrolled. This is now fixed.

Finally you can pan to close the bottom sheet. However if you have scrolled down, then you have to scroll up, lift your finger, then pan again. It's annoying but will have to be fixed later.

I tried alternative bottom-sheets and they either had different issues or didn't work for the web export.