refactor: Prep auth for multiple accounts (#1943)
The previous flow only worked for accounts with a single cross save account. For accounts where the user needs to select the Destiny 2 account the system would break.

Further changes were needed as the process is async. A list of accounts needs to be shown and then the user selects the correct one.

This was done in an over 'clever' way in Ishtar and GG was heading down the same path. This makes the login more complex, as in the total number of functions. But removes the 'clever' part.