fix: tabs were hard to tap (#2093)
This was due to multiple issues. The marginBottom clips the tappable area making it smaller. This is now removed.

This had been added to centre the text. The text is now correctly centred by overriding part of the style.

Finally the pressable is replaced with a TouchableOpacity that visually responds instantly to a tap to further give feedback on a successful press.