refactor: Optimize updateAllPages
This seem to reduce the time by ~40 when it reruns for things like changing the sort. 

The change mostly comes from simply not doing work on pages that have not changes. A small change comes from an error where the lostItems count would always be updated. Now the values are deepEqual checked.