refactor: Improve start up
This change creates a new state variable 'appReady'. This is true when the state has hydrated and both the custom item defintion and bungie definitions have all successfully been downloaded or on app restart loaded.

The MainDrawer UI will not be shown till the app is ready making it impossible for the auto refresh or manual refresh to be activated.

The splash screen is then only hidden when the app is ready. And exception to this is when the app is not authenticated. In this case the splash is dimissed and then the login screen is shown.