refactor: simplify focus and page data (#2125)
Instead of a non standard getData function this simply binds the page to a new getPageData() function in Zustand.

Also onFocus is replaced with useFocusEffect. The benefit being useFocusEffect only runs when the page is focused. isFocused was running even when a page lost focus and triggering a re-render of a view you no longer could see.